[1.0.0-beta] first beta snapshot release
This commit contains the changes for the first major beta release.
I know that there is a lot of code here, but the documentation for it is still work in progress.
The general usage is still as it was in alpha, but there are a lot more functionalities introduced.
If you feel like anything is missing for a first stable release, feel free to leave feedback in the feedback form, or to create a feature-request issue.