feature(collections): pagination support
Added support for pagination that can be reused in multiple modules like web and database.
The goal of this is to make the abstraction once and have the potential for shared mapping logic.
Also included is an autoconfiguration for the jackson integration to make serialization of this new feature smooth.